feat 【角色管理】维护角色与用户的关系
  - 给角色分配用户
  - 查询某个角色下的用户列表（以部门组织架构作树形结构展示）

+++ src/main/java/com/pyr/permission/domain/user/service/SysUserService.java
@@ -67,7 +67,7 @@ public class SysUserService {
         return sysUserMapper.findByKeyword(keyword);
     }
 
-    public SysUser findById(Integer userId) {
+    public SysUser findById(Long userId) {
         return sysUserMapper.selectById(userId);
     }
